+++ perun/utils/__init__.py
@@ -303,7 +303,7 @@ def get_supported_module_names(package):
         'collect': ['trace', 'memory', 'time', 'complexity', 'bounds'],
         'postprocess': ['clusterizer', 'normalizer', 'regression-analysis',
                         'regressogram', 'moving-average', 'kernel-regression'],
-        'view': ['bars', 'flamegraph', 'flow', 'heapmap', 'raw', 'scatter']
+        'view': ['bars', 'flamegraph', 'flow', 'heapmap', 'raw', 'scatter', 'table']
     }[package]